Move utilities file out of calculations directory
