+++ client/src/app/calculations/geography.ts
@@ -3,7 +3,7 @@ import type {
   RegionState,
   StateState,
 } from '@/app/redux/reducers/geography';
-import { sortObjectByKeys } from '@/app/calculations/utilities';
+import { sortObjectByKeys } from '@/app/utilities';
 import type { RegionId, RegionName, StateId } from '@/app/config';
 /**
  * Excel: "CountyFIPS" sheet.
